+++ htdocs/google/ChangeLog.md
@@ -4,7 +4,7 @@
 ## 6.8
 
 - Add more information on setup page
-- NEW Add job position and company in synch of contacts
+- NEW Add job position and company in sync of contacts, add also in mass init sync.
 
 
 ## 6.7